services/ai-document-quality/pyproject.toml
[project]
name = "transport"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=23.0.0",
    "isort>=6.0.0",
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "pandas>=2.0.0",
    "pdf2image>=1.17.0",
    "pytesseract>=0.3.13",
    "pytz>=2024.2",
    "requests>=2.32.3",
    "torchvision>=0.16.0",
    "tqdm>=4.66.0",
    "xlsxwriter>=3.2.2",
    "pillow>=10.0.0",
    "scikit-learn>=1.3.0",
    "flask>=3.0.0",
    "pyyaml>=6.0",
    "uvicorn[standard]>=0.24.0",
    "torchinfo>=1.8.0",
    "mlflow>=2.8.0",
    "torch>=2.1.0",
    "types-tqdm>=4.67.0.20250516",
    "seaborn>=0.12.0",
    "tensorboard>=2.14.0",
    "albumentations>=1.3.0",
    "fastapi>=0.104.0",
    "python-multipart>=0.0.6",
    "gradio>=4.0.0",
    "streamlit>=1.28.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "python-dotenv>=1.0.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
    "ipywidgets>=8.1.0",
]
